Serialize now uses u8s instead of usizes
usizes break across different machines, and u8s are more flexible. Code
is also a lot cleaner now

+++ tests/src/lib.rs
@@ -4,11 +4,11 @@ use instrs_core::*;
 #[derive(Serialize)]
 pub enum Instruction {
     Add {
-        a: usize,
-        b: usize,
-        addr: usize,
+        a: u8,
+        b: u8,
+        addr: u8,
     },
-    Jump(usize),
+    Jump(u32),
     Halt,
 }
